# backend/docs/swagger.yaml
definitions:
...
  api.ExamResponse:
    properties:
      exam_id:
        type: integer
      is_unlocked:
        type: boolean
      time_limit_minutes:
        type: integer
      title:
        type: string
    type: object
...
  api.Response:
    properties:
      data: {}
      error:
        type: string
      language:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  api.createExamRequest:
    properties:
      is_unlocked:
        type: boolean
      time_limit_minutes:
        minimum: 1
        type: integer
      title:
        type: string
    required:
    - time_limit_minutes
    - title
    type: object
...
  api.updateExamRequest:
    properties:
      is_unlocked:
        type: boolean
      time_limit_minutes:
        minimum: 1
        type: integer
      title:
        type: string
    type: object
...
  monitoring.Alert:
    properties:
      count:
        description: Number of times this alert fired
        type: integer
      details:
        additionalProperties: true
        type: object
      id:
        type: string
      last_fired:
        type: string
      level:
        $ref: '#/definitions/monitoring.AlertLevel'
      message:
        type: string
      name:
        type: string
      resolved:
        type: boolean
      resolved_at:
        type: string
      source:
        type: string
      timestamp:
        type: string
    type: object
  monitoring.AlertLevel:
    enum:
    - INFO
    - WARNING
    - CRITICAL
    type: string
    x-enum-varnames:
    - AlertLevelInfo
    - AlertLevelWarning
    - AlertLevelCritical
...
info:
  contact: {}
paths:
...
  /alerts/history:
    get:
      consumes:
      - application/json
      description: Get recent alert history
      parameters:
      - default: 50
        description: Maximum number of alerts to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Alert history retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/monitoring.Alert'
                  type: array
              type: object
      summary: Get alert history
      tags:
      - monitoring
...
  /api/v1/exams:
    get:
      consumes:
      - application/json
      description: Get a list of all exams with pagination
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Exams retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api.ExamResponse'
                  type: array
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Failed to retrieve exams
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: List exams
      tags:
      - exams
    post:
      consumes:
      - application/json
      description: Add a new exam to the database
      parameters:
      - description: Exam object to create
        in: body
        name: exam
        required: true
        schema:
          $ref: '#/definitions/api.createExamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Exam created successfully
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.ExamResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Failed to create exam
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new exam
      tags:
      - exams
  /api/v1/exams/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific exam by its ID
      parameters:
      - description: Exam ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Exam deleted successfully
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Invalid exam ID
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Failed to delete exam
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete an exam
      tags:
      - exams
    get:
      consumes:
      - application/json
      description: Retrieve a specific exam entry by its ID
      parameters:
      - description: Exam ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Exam retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.ExamResponse'
              type: object
        "400":
          description: Invalid exam ID
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Exam not found
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Failed to retrieve exam
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Get an exam by ID
      tags:
      - exams
    put:
      consumes:
      - application/json
      description: Update an existing exam by ID
      parameters:
      - description: Exam ID
        in: path
        name: id
        required: true
        type: integer
      - description: Exam fields to update
        in: body
        name: exam
        required: true
        schema:
          $ref: '#/definitions/api.updateExamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Exam updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.ExamResponse'
              type: object
        "400":
          description: Invalid request body or exam ID
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Exam not found
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Failed to update exam
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Update an exam
      tags:
      - exams
...
swagger: "2.0"